get-new-image names smarter

+++ 0-prep-images/get_newimagenames.py
@@ -8,16 +8,23 @@ def main():
     ofilepath = params["outputdir"] + "meta-data/" + "new-images.txt"
     PATH = params["newimages"]
     OPATH = params["images"]
-    ofile = open (ofilepath, "w")
 
-    tifs = [y for x in os.walk(PATH) for y in glob(os.path.join(x[0], '*.tif'))]
-    jpgs = [y for x in os.walk(PATH) for y in glob(os.path.join(x[0], '*.jpg'))]
-    result = tifs + jpgs 
+    if os.path.exists(params["newimages"]):
+        print ("moving all the images to main folder")
+        ofile = open (ofilepath, "w")
 
-    for r in result:
-        print (r, file=ofile)
-        fname = r.split('/')[-1]
-        os.rename(r, OPATH + fname)
+        tifs = [y for x in os.walk(PATH) for y in glob(os.path.join(x[0], '*.tif'))]
+        jpgs = [y for x in os.walk(PATH) for y in glob(os.path.join(x[0], '*.jpg'))]
+        result = tifs + jpgs 
+
+        for r in result:
+            print (r, file=ofile)
+            fname = r.split('/')[-1]
+            os.rename(r, OPATH + fname)
+
+        os.rmdir(params["newimages"])
+    else:
+        print ("images already in proper place, ready to procede")
 
 if __name__=="__main__":
     main()
